backend_django/securePassKeeper/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.5
Django==4.2.7
django-cors-headers==4.3.1
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
install==1.3.5
packaging==23.2
phonenumbers==8.13.26
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
